+++ internal/service/returns.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"errors"
 	"time"
 
 	"gitlab.ozon.dev/yuweebix/homework-1/internal/models"
@@ -34,11 +35,22 @@ func (s Service) AcceptReturn(o *models.Order) (err error) {
 	ro.Status = models.StatusReturned
 	ro.Hash = hash.GenerateHash() // HASH
 
-	err = s.storage.DeleteOrder(ro)
+	err = s.storage.UpdateOrder(ro)
 	if err != nil {
+		if errors.Is(err, errors.ErrUnsupported) {
+			err = s.storage.DeleteOrder(ro)
+			if err != nil {
+				return err
+			}
+
+			err = s.storage.CreateOrder(ro)
+			if err != nil {
+				return err
+			}
+		}
 		return err
 	}
-	return s.storage.CreateOrder(ro)
+	return nil
 }
 
 // ListReturns выводит список заказов
